refactor(client): removed relative imports and restructured
