+++ client/agents/sql_agent/agents.py
@@ -1,11 +1,9 @@
-from .tools import sql_system_prompt
+from client.agents.common.base import Agent
+from client.agents.sql_agent.models import SQLResponse
+from client.agents.sql_agent.prompts import sql_system_prompt
 
-from ..common.base import Agent
-from .models import SQLResponse
-
-
-SQL_Agent = Agent(
-    name="SQL_Agent",
+sql_agent = Agent(
+    name="sql_agent",
     instructions=sql_system_prompt,
     response_format=SQLResponse,
 )
