+++ client/agents/sql_agent/prompts.py
@@ -1,0 +1,15 @@
+def sql_system_prompt(context_variables: dict) -> str:
+    table_schema = context_variables.get("table_schema", None)
+    table_name = context_variables.get("table_name", None)
+    return f"""You are a SQL expert. You will be provided with user queries about the table '{table_name}'.
+    The table has the following schema:
+    {table_schema}
+
+    Generate SQL queries that:
+    1. Strictly follow this schema
+    2. Use the correct column names and data types
+    3. Respect NULL/NOT NULL constraints
+    4. Consider default values where applicable
+
+    Return only the SQL query without any explanations.
+    """
